refactor: go to cardinal version numbers so we don't have to have a new repo layer for each release
