+++ components/crypto/crypto-persistence/src/main/kotlin/net/corda/crypto/persistence/v2schema/V2CryptoRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package net.corda.crypto.persistence.v50ga
+package net.corda.crypto.persistence.v2schema
 
 import net.corda.crypto.persistence.CryptoRepository
 import net.corda.crypto.persistence.WrappingKeyInfo
@@ -6,14 +6,13 @@ import net.corda.orm.utils.transaction
 import net.corda.orm.utils.use
 import java.time.Instant
 import java.time.LocalDateTime
-import java.util.*
 import javax.persistence.EntityManagerFactory
 
-class V50GACryptoRepositoryImpl(private val entityManagerFactory: () -> EntityManagerFactory) : CryptoRepository {
+class V2CryptoRepositoryImpl(private val entityManagerFactory: () -> EntityManagerFactory) : CryptoRepository {
     override fun saveWrappingKey(alias: String, key: WrappingKeyInfo) {
         entityManagerFactory().transaction { em ->
             em.persist(
-                V50GAWrappingKeyEntity(
+                V2WrappingKeyEntity(
                     alias = alias,
                     created = Instant.now(),
                     encodingVersion = key.encodingVersion,
@@ -26,7 +25,7 @@ class V50GACryptoRepositoryImpl(private val entityManagerFactory: () -> EntityMa
     }
 
     override fun findWrappingKey(alias: String): WrappingKeyInfo? = entityManagerFactory().use { em ->
-        em.find(V50GAWrappingKeyEntity::class.java, alias)?.let { rec ->
+        em.find(V2WrappingKeyEntity::class.java, alias)?.let { rec ->
             WrappingKeyInfo(
                 encodingVersion = rec.encodingVersion,
                 algorithmName = rec.algorithmName,
